Split up vector lists classes into own files
