+++ lib/src/vector_math_lists/vector_list.dart
@@ -126,72 +126,3 @@ abstract class VectorList<T extends Vector> {
     store(index, v);
   }
 }
-
-/// A list of [Vector2].
-class Vector2List extends VectorList<Vector2> {
-  /// Create a new vector list with [length] elements. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector2List(int length, [int offset = 0, int stride = 0])
-      : super(length, 2, offset, stride);
-
-  /// Create a new vector list from a list of vectors. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector2List.fromList(List<Vector2> list, [int offset = 0, int stride = 0])
-      : super.fromList(list, 2, offset, stride);
-
-  /// Create a new vector list as a view of [buffer]. Optionally it is possible
-  /// to specify a [offset] in the [buffer] and a [stride] between each vector.
-  Vector2List.view(Float32List buffer, [int offset = 0, int stride = 0])
-      : super.view(buffer, 2, offset, stride);
-
-  @override
-  Vector2 newVector() {
-    return new Vector2.zero();
-  }
-}
-
-/// A list of [Vector3].
-class Vector3List extends VectorList<Vector3> {
-  /// Create a new vector list with [length] elements. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector3List(int length, [int offset = 0, int stride = 0])
-      : super(length, 3, offset, stride);
-
-  /// Create a new vector list from a list of vectors. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector3List.fromList(List<Vector3> list, [int offset = 0, int stride = 0])
-      : super.fromList(list, 3, offset, stride);
-
-  /// Create a new vector list as a view of [buffer]. Optionally it is possible
-  /// to specify a [offset] in the [buffer] and a [stride] between each vector.
-  Vector3List.view(Float32List buffer, [int offset = 0, int stride = 0])
-      : super.view(buffer, 3, offset, stride);
-
-  @override
-  Vector3 newVector() {
-    return new Vector3.zero();
-  }
-}
-
-/// A list of [Vector4].
-class Vector4List extends VectorList<Vector4> {
-  /// Create a new vector list with [length] elements. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector4List(int length, [int offset = 0, int stride = 0])
-      : super(length, 4, offset, stride);
-
-  /// Create a new vector list from a list of vectors. Optionally it is possible
-  /// to specify an [offset] in the [buffer] and a [stride] between each vector.
-  Vector4List.fromList(List<Vector4> list, [int offset = 0, int stride = 0])
-      : super.fromList(list, 4, offset, stride);
-
-  /// Create a new vector list as a view of [buffer]. Optionally it is possible
-  /// to specify a [offset] in the [buffer] and a [stride] between each vector.
-  Vector4List.view(Float32List buffer, [int offset = 0, int stride = 0])
-      : super.view(buffer, 4, offset, stride);
-
-  @override
-  Vector4 newVector() {
-    return new Vector4.zero();
-  }
-}
