+++ lib/vector_math_lists.dart
@@ -3,11 +3,13 @@
 // license that can be found in the LICENSE file.
 
 /// A library for working with lists of vectors in a memory efficient way.
-
 library vector_math_lists;
 
 import 'dart:typed_data';
 import 'dart:math' as Math;
 import 'package:vector_math/vector_math.dart';
 
-part 'src/vector_math_lists/vector.dart';
+part 'src/vector_math_lists/vector_list.dart';
+part 'src/vector_math_lists/vector2_list.dart';
+part 'src/vector_math_lists/vector3_list.dart';
+part 'src/vector_math_lists/vector4_list.dart';
